SocreListContent を分離
